pyproject.toml
# ===============================
# Machine_Learning-based_Quantitative_Trading_Strategies
# 機械学習を用いたクオンツ戦略：研究／実装用プロジェクト設定（uv対応）
# ※ 依存関係は uv で解決し、`uv sync` → `uv run <script>` を基本運用とする
# ===============================

[project]
# プロジェクトの基本情報
name = "ml-quant-strategies"
version = "0.1.0"
description = "機械学習を用いたクオンツ（数量的）投資戦略の研究・実装"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11,<3.12"

# 研究・実装のコア依存関係
# - データ処理／I/O：pandas, numpy, pyarrow, pyyaml
# - 可視化：matplotlib, plotly
# - 機械学習：scikit-learn, torch（CPU版を標準。GPUは各自環境で別途インストール推奨）
# - 市場データ取得：yfinance
# - ノートブック：jupyter
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "pyarrow>=15",
  "pyyaml>=6.0.2",
  "matplotlib>=3.8",
  "plotly>=5.20",
  "scikit-learn>=1.4",
  "torch>=2.2; platform_system != 'Windows'",        # Linux/MacのCPU版（GPU/CUDAは各自で）
  "torch==2.2.2; platform_system == 'Windows'",      # Windowsの例（必要に応じて調整）
  "yfinance>=0.2.50",
  "jupyter>=1.0",
  # バックテスト系（任意）：backtesting.py はPandasベース
  "backtesting>=0.3.3",
]

# プロジェクトURL（任意）：レポジトリやドキュメント等に置き換え
[project.urls]
homepage = "https://github.com/<your-org-or-name>/<your-repo>"
repository = "https://github.com/<your-org-or-name>/<your-repo>"

# 追加の開発用依存（`uv add --dev` で解決される。CIやローカル品質管理向け）
[tool.uv]
# scripts エイリアスは本環境の uv では未対応のため、明示的に `uv run python ...` を使用する
# 例:
#   uv run python scripts/fetch_daily.py
#   uv run python -m app.main

# -------- オプションのエクストラ依存関係（必要な人だけ `uv add '.[viz]'` のように使用） --------
[project.optional-dependencies]
# 可視化・報告周りを強化したい場合
viz = [
  "seaborn>=0.13",    # ※ 本プロジェクトではデフォルトはmatplotlib/plotly。好みで追加
]
# 強化学習（将来予定）：必要になったときに個別導入推奨（環境依存が強いため）
rl = [
  "stable-baselines3>=2.3",
]

# -------- Lint/Format 等（導入する場合。未導入なら削除可） --------
[tool.ruff]
# コードスタイル：プロジェクト事情に合わせて調整
line-length = 100
target-version = "py311"
extend-select = ["I"]   # import順序の自動整形など

[tool.pytest.ini_options]
# PyTestの基本設定（必要に応じて編集）
addopts = "-q"
testpaths = ["tests"]

# ===============================
# 使い方メモ（READMEにも記載推奨）
# 1) uv のインストール（初回のみ）
#    pipx install uv
# 2) 依存解決
#    uv sync
# 3) データ取得
#    uv run fetch
# 4) 動作確認
#    uv run run
# ===============================
